Add unit conversion for coil length calculation
The code changes include adding a unit conversion feature for the calculation of coil length. The `compute_coil_length` function now checks the value of the `unit_var` variable and adjusts the length calculation accordingly. If the unit is set to 'I', the length is divided by 12 to convert it to inches. Otherwise, if the unit is not 'I', the length is divided by 1000 to convert it to meters.

This commit message follows the established convention of using a prefix to indicate the type of change, followed by a concise and descriptive message.